Make it run as a binary

+++ src/config.rs
@@ -2,7 +2,7 @@
 
 pub struct Config {
     pub filename: String, // $1
-    pub tilesize: u8, // 256
+    pub tilesize: u32, // 256
     pub zoomlevel: u8, // eg 0 - 5
     // don't need MGK
     pub folder: String, // $OUTDIR out
